==> notebooks/3-Basic_Models.ipynb <==
# 3 - Models<p>
I'm doing an initial run with these regression models to see how well they perform:<ul>
Linear Regression<br>
Support Vector Regressor<br>
Random Forest Regressor<br>
XGBoost Regressor

Import pickled data and libraries:

import pandas as pd
dc_data = pd.read_pickle('../pickles/dc_data.pkl')
london_data = pd.read_pickle('../pickles/london_data.pkl')
from sklearn.metrics import r2_score, root_mean_squared_error
from sklearn.linear_model import LinearRegression
from sklearn.svm import SVR
from sklearn.ensemble import RandomForestRegressor
from xgboost import XGBRegressor

Split the data into trains and tests, but not randomly, because this data being ordered by time matters.

train_dc = dc_data[:int(0.8*(len(dc_data)))]
X_train_dc = train_dc.drop('count',axis=1)
y_train_dc = train_dc['count']

test_dc = dc_data[int(0.8*(len(dc_data))):]
X_test_dc = test_dc.drop('count',axis=1)
y_test_dc = test_dc['count']

X_train_dc.to_pickle('../pickles/split/X_train_dc.pkl')
y_train_dc.to_pickle('../pickles/split/y_train_dc.pkl')
X_test_dc.to_pickle('../pickles/split/X_test_dc.pkl')
y_test_dc.to_pickle('../pickles/split/y_test_dc.pkl')

train_l = london_data[:int(0.8*(len(london_data)))]
X_train_l = train_l.drop('count',axis=1)
y_train_l = train_l['count']

test_l = london_data[int(0.8*(len(london_data))):]
X_test_l = test_l.drop('count',axis=1)
y_test_l = test_l['count']

X_train_l.to_pickle('../pickles/split/X_train_l.pkl')
y_train_l.to_pickle('../pickles/split/y_train_l.pkl')
X_test_l.to_pickle('../pickles/split/X_test_l.pkl')
y_test_l.to_pickle('../pickles/split/y_test_l.pkl')

### Linear Regression

linreg_dc = LinearRegression()
linreg_dc.fit(X_train_dc, y_train_dc)

pred_y_train_dc = linreg_dc.predict(X_train_dc)
pred_y_test_dc = linreg_dc.predict(X_test_dc)

dc_train_rmse = root_mean_squared_error(y_train_dc, pred_y_train_dc)
dc_train_r2 = r2_score(y_train_dc, pred_y_train_dc)
dc_test_rmse = root_mean_squared_error(y_test_dc, pred_y_test_dc)
dc_test_r2 = r2_score(y_test_dc, pred_y_test_dc)

print('DC dataset results \nTrain set:')
print(f"Root Mean Squared Error: {dc_train_rmse}")
print(f"R-squared: {dc_train_r2}")
print('\nTest set:')
print(f"Root Mean Squared Error: {dc_test_rmse}")
print(f"R-squared: {dc_test_r2}")

linreg_l = LinearRegression()
linreg_l.fit(X_train_l, y_train_l)

pred_y_train_l = linreg_l.predict(X_train_l)
pred_y_test_l = linreg_l.predict(X_test_l)

l_train_rmse = root_mean_squared_error(y_train_l, pred_y_train_l)
l_train_r2 = r2_score(y_train_l, pred_y_train_l)
l_test_rmse = root_mean_squared_error(y_test_l, pred_y_test_l)
l_test_r2 = r2_score(y_test_l, pred_y_test_l)

print('London dataset results \nTrain set:')
print(f"Root Mean Squared Error: {l_train_rmse}")
print(f"R-squared: {l_train_r2}")
print('\nTest set:')
print(f"Root Mean Squared Error: {l_test_rmse}")
print(f"R-squared: {l_test_r2}")

### SVR

svr_dc = SVR()
svr_dc.fit(X_train_dc, y_train_dc)

pred_svr_dc_train = svr_dc.predict(X_train_dc)
pred_svr_dc_test = svr_dc.predict(X_test_dc)

svr_dc_rmse_train = root_mean_squared_error(y_train_dc, pred_svr_dc_train)
svr_dc_r2_train = r2_score(y_train_dc, pred_svr_dc_train)

svr_dc_rmse_test = root_mean_squared_error(y_test_dc, pred_svr_dc_test)
svr_dc_r2_test = r2_score(y_test_dc, pred_svr_dc_test)

print('DC dataset results \nTrain set:')
print(f"Root Mean Squared Error: {svr_dc_rmse_train}")
print(f"R-squared: {svr_dc_r2_train}")
print('\nTest set:')
print(f"Root Mean Squared Error: {svr_dc_rmse_test}")
print(f"R-squared: {svr_dc_r2_test}")

svr_l = SVR()
svr_l.fit(X_train_l, y_train_l)

pred_svr_l_train = svr_l.predict(X_train_l)
pred_svr_l_test = svr_l.predict(X_test_l)

svr_l_rmse_train = root_mean_squared_error(y_train_l, pred_svr_l_train)
svr_l_r2_train = r2_score(y_train_l, pred_svr_l_train)

svr_l_rmse_test = root_mean_squared_error(y_test_l, pred_svr_l_test)
svr_l_r2_test = r2_score(y_test_l, pred_svr_l_test)

print('London dataset results \nTrain set:')
print(f"Root Mean Squared Error: {svr_l_rmse_train}")
print(f"R-squared: {svr_l_r2_train}")
print('\nTest set:')
print(f"Root Mean Squared Error: {svr_l_rmse_test}")
print(f"R-squared: {svr_l_r2_test}")

### Random Forest Regressor

forest_dc = RandomForestRegressor(random_state=42)
forest_dc.fit(X_train_dc, y_train_dc)

forest_pred_dc_train = forest_dc.predict(X_train_dc)
forest_pred_dc_test = forest_dc.predict(X_test_dc)

forest_dc_rmse_train = root_mean_squared_error(y_train_dc, forest_pred_dc_train)
forest_dc_r2_train = r2_score(y_train_dc, forest_pred_dc_train)

forest_dc_rmse_test = root_mean_squared_error(y_test_dc, forest_pred_dc_test)
forest_dc_r2_test = r2_score(y_test_dc, forest_pred_dc_test)

print('DC dataset results \nTrain set:')
print(f"Root Mean Squared Error: {forest_dc_rmse_train}")
print(f"R-squared: {forest_dc_r2_train}")
print('\nTest set:')
print(f"Root Mean Squared Error: {forest_dc_rmse_test}")
print(f"R-squared: {forest_dc_r2_test}")

forest_l = RandomForestRegressor(random_state=42)
forest_l.fit(X_train_l, y_train_l)

forest_pred_l_train = forest_l.predict(X_train_l)
forest_pred_l_test = forest_l.predict(X_test_l)

forest_l_rmse_train = root_mean_squared_error(y_train_l, forest_pred_l_train)
forest_l_r2_train = r2_score(y_train_l, forest_pred_l_train)

forest_l_rmse_test = root_mean_squared_error(y_test_l, forest_pred_l_test)
forest_l_r2_test = r2_score(y_test_l, forest_pred_l_test)

print('London dataset results \nTrain set:')
print(f"Root Mean Squared Error: {forest_l_rmse_train}")
print(f"R-squared: {forest_l_r2_train}")
print('\nTest set:')
print(f"Root Mean Squared Error: {forest_l_rmse_test}")
print(f"R-squared: {forest_l_r2_test}")

### XGBoost Regressor

xgbr_dc = XGBRegressor(objective='reg:squarederror', random_state=42)
xgbr_dc.fit(X_train_dc, y_train_dc)

xgbr_pred_dc_train = xgbr_dc.predict(X_train_dc)
xgbr_pred_dc_test = xgbr_dc.predict(X_test_dc)

xgbr_dc_rmse_train = root_mean_squared_error(y_train_dc, xgbr_pred_dc_train)
xgbr_dc_r2_train = r2_score(y_train_dc, xgbr_pred_dc_train)

xgbr_dc_rmse_test = root_mean_squared_error(y_test_dc, xgbr_pred_dc_test)
xgbr_dc_r2_test = r2_score(y_test_dc, xgbr_pred_dc_test)

print(f"DC dataset results: \nRoot Mean Squared Error: \n  Train: {xgbr_dc_rmse_train} \n  Test: {xgbr_dc_rmse_test}")
print(f"\nR2: \n  Train: {xgbr_dc_r2_train} \n  Test: {xgbr_dc_r2_test}")

xgbr_l = XGBRegressor(objective='reg:squarederror', random_state=42)
xgbr_l.fit(X_train_l, y_train_l)

xgbr_pred_l_train = xgbr_l.predict(X_train_l)
xgbr_pred_l_test = xgbr_l.predict(X_test_l)

xgbr_l_rmse_train = root_mean_squared_error(y_train_l, xgbr_pred_l_train)
xgbr_l_r2_train = r2_score(y_train_l, xgbr_pred_l_train)

xgbr_l_rmse_test = root_mean_squared_error(y_test_l, xgbr_pred_l_test)
xgbr_l_r2_test = r2_score(y_test_l, xgbr_pred_l_test)

print(f"London dataset results: \nRoot Mean Squared Error: \n  Train: {xgbr_l_rmse_train} \n  Test: {xgbr_l_rmse_test}")
print(f"\nR2: \n  Train: {xgbr_l_r2_train} \n  Test: {xgbr_l_r2_test}")

Compare the models' performance against each other:

print(f"Linear Regression results: \n  DC RMSE: {dc_test_rmse} \n  DC R2: {dc_test_r2} \n  London RMSE: {l_test_rmse} \n  London R2: {l_test_r2}")
print(f"\nSVR results: \n  DC RMSE: {svr_dc_rmse_test} \n  DC R2: {svr_dc_r2_test} \n  London RMSE: {svr_l_rmse_test} \n  London R2: {svr_l_r2_test}")
print(f"\nRandom Forest results: \n  DC RMSE: {forest_dc_rmse_test} \n  DC R2: {forest_dc_r2_test} \n  London RMSE: {forest_l_rmse_test} \n  London R2: {forest_l_r2_test}")
print(f"\nXGBRegressor results: \n  DC RMSE: {xgbr_dc_rmse_test} \n  DC R2: {xgbr_dc_r2_test} \n  London RMSE: {xgbr_l_rmse_test} \n  London R2: {xgbr_l_r2_test}")

Of these four, the best performers, based on both the RMSE **and** the R2 are `RandomForestRegressor` and `XGBRegressor`. I'm going to take these two forward to hyperparameter tuning.
